Fix an issue where compilation may fail on some non-x64 platforms

+++ src/codec/packet/side_data.rs
@@ -195,6 +195,6 @@ impl<'a> SideData<'a> {
     }
 
     pub fn data(&self) -> &[u8] {
-        unsafe { slice::from_raw_parts((*self.as_ptr()).data, (*self.as_ptr()).size) }
+        unsafe { slice::from_raw_parts((*self.as_ptr()).data, (*self.as_ptr()).size as usize) }
     }
 }
